2007-03-07 Erik Hjortsberg  <erik.hjortsberg@iteam.se>
	* src/components/ogre/Makefile.am, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/Makefile.am, src/components/ogre/carpenter/Makefile.am, src/components/ogre/environment/Makefile.am, src/components/ogre/environment/meshtree/Makefile.am, src/components/ogre/input/Makefile.am, src/components/ogre/jesus/Makefile.am, src/components/ogre/manipulation/Makefile.am, src/components/ogre/model/Makefile.am, src/components/ogre/ogreopcode/Makefile.am, src/components/ogre/ogreopcode/opcode/Makefile.am, src/components/ogre/scripting/Makefile.am, src/components/ogre/scripting/bindings/lua/Makefile.am, src/components/ogre/scripting/bindings/lua/helpers/Makefile.am, src/components/ogre/scripting/bindings/lua/ogre/Makefile.am, src/components/ogre/sound/Makefile.am, src/components/ogre/terrain/Makefile.am, src/components/ogre/widgets/Makefile.am, src/extensions/Makefile.am, src/framework/Makefile.am, src/framework/bindings/lua/Makefile.am, src/framework/bindings/lua/eris/Makefile.am, src/services/Makefile.am, src/services/bindings/lua/Makefile.am, src/services/config/Makefile.am, src/services/logging/Makefile.am, src/services/metaserver/Makefile.am, src/services/scripting/Makefile.am, src/services/server/Makefile.am, src/services/sound/Makefile.am: better structure of the sublibs
	* src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeData2DManager.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeOctree.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeSceneManager.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeTexture.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeTexture_InstantBaseTextureShadowed.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/filetutils.cpp: remove annoying warnings

+++ src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeData2DManager.cpp
@@ -1793,7 +1793,7 @@ namespace Ogre
                 unsigned int Lod = 0;
                 if (t && t->isLoaded())
                     Lod = t->getRenderable()->getRenderLevel();
-                const Vector3 normalVector = getNormalAt(pageX, pageZ, x, z, Lod);
+                const Vector3 normalVector = getNormalAt(pageX, pageZ, static_cast<unsigned int>(x), static_cast<unsigned int>(z), Lod);
                 const Real slope = 1.0f + normalVector.y;
                 if (slope < 0.001f)
                     return 0.0f;
